Refactor torch util functions
