+++ src/garage/torch/algos/ddpg.py
@@ -9,7 +9,7 @@ import torch
 from garage import _Default, make_optimizer
 from garage.np.algos.off_policy_rl_algorithm import OffPolicyRLAlgorithm
 from garage.replay_buffer import PathBuffer
-import garage.torch.utils as tu
+from garage.torch import dict_np_to_torch, torch_to_np
 
 
 class DDPG(OffPolicyRLAlgorithm):
@@ -184,7 +184,7 @@ class DDPG(OffPolicyRLAlgorithm):
                         self.buffer_batch_size)
                 else:
                     samples = self.replay_buffer.sample(self.buffer_batch_size)
-                qf_loss, y, q, policy_loss = tu.torch_to_np(
+                qf_loss, y, q, policy_loss = torch_to_np(
                     self.optimize_policy(samples))
 
                 self._episode_policy_losses.append(policy_loss)
@@ -235,7 +235,7 @@ class DDPG(OffPolicyRLAlgorithm):
             qval: Q-value predicted by the Q-network.
 
         """
-        transitions = tu.dict_np_to_torch(samples_data)
+        transitions = dict_np_to_torch(samples_data)
 
         # pylint: disable=fixme
         # TODO: remove this check once HER uses PathBuffer.
